Deflate and Convert Comexstat Data

This function deflates USD-denominated values in Comexstat data (trade statistics) using specified deflators (CPI for USD, IPCA for BRL after converting using the exchange rate of the date of the statistics)

Usage

comex_deflate(data, basedate = NULL, deflators = get_deflators(na_omit = TRUE))

Arguments

data

A data frame or tibble containing Comexstat data.

basedate

An optional date object specifying the base date for deflation. If NULL, the latest available date in the deflators data is used.

deflators

A data frame containing deflator time series data, including columns cpi (Consumer Price Index), ipca_i (Brazilian IPCA index), and date. Defaults to the get_deflators() function's output with missing values removed.

Value

A modified version of the input data, with the following changes:

New columns for deflated USD values (e.g., fob_usd_deflated, cif_usd_deflated).

New columns for BRL values based on the exchange rate and USD values (e.g., fob_brl, cif_brl).

New columns for deflated BRL values using the IPCA deflator (e.g., fob_brl_deflated).

Details

The function performs the following steps:

Handles missing values in deflators, issuing a warning if present.

Determines the base date for deflation, either from basedate input or the latest date in deflators.

Calculates deflation ratios (cpi_r, ipca_r) for each date relative to the base date.

Joins the deflators data with the input data based on date.

Deflates the USD-denominated columns (fob_usd, cif_usd, etc.) using cpi_r.

Converts USD values to BRL based on the exchange rate in the brlusd column.

Deflates the BRL values using ipca_r.

Arranges the resulting data by date.

Examples

if (FALSE) {
# Deflate using default data and latest base date
deflated_data <- comex_deflate()

# Deflate with a specific base date (e.g., 2023-12-31)
deflated_data <- comex_deflate(basedate = as.Date('2023-12-31'))
}