comexr: Client for the Brazilian Foreign Trade Statistics API ('ComexStat')

Interface to the 'ComexStat' API https://comexstat.mdic.gov.br/ from the Brazilian Ministry of Development, Industry, Trade and Services (MDIC). Provides access to detailed export and import data, including general trade statistics (1997-present), city-level data, historical data (1989-1996), and auxiliary tables with product codes (NCM - Nomenclatura Comum do Mercosul, NBM - Nomenclatura Brasileira de Mercadorias, HS - Harmonized System), countries, economic classifications (CGCE - Classificacao por Grandes Categorias Economicas, SITC - Standard International Trade Classification, ISIC - International Standard Industrial Classification), and other categories. Uses only 'httr2' for HTTP requests and 'cli' for console messages.

See also

Author

Maintainer: Andre Leite leite@castlab.org

Authors:

Marcos Wasilew marcos.wasilew@gmail.com

Hugo Vasconcelos hugo.vasconcelos@ufpe.br

Carlos Amorin carlos.agaf@ufpe.br

Diogo Bezerra diogo.bezerra@ufpe.br